tests/wpt/web-platform-tests/speculation-rules/prerender/restriction-web-nfc.https.html

<!DOCTYPE html>
<title>Access to the Web NFC API is deferred</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const uid = token();
  const bc = new PrerenderChannel('test-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  const url = `resources/web-nfc.https.html?uid=${uid}`;
  window.open(url, '_blank', 'noopener');

  const result = await gotMessage;
  const expected = [
    {event: 'started waiting NDEFReader.[write|scan]',  prerendering: true},
    {event: 'prerendering change',                      prerendering: false},
    {event: 'ndef.write() failed',                      prerendering: false},
    {event: 'ndef.scan() failed',                       prerendering: false},
    {event: 'finished waiting NDEFReader.[write|scan]', prerendering: false},
  ];
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event[${i}]`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering[${i}]`);
  }

  // Send a close signal to PrerenderEventCollector on the prerendered page.
  new PrerenderChannel('close', uid).postMessage('');
}, `The access to the Web NFC API should be deferred until the prerendered
    page is activated`);

</script>
</body>
